First steps in allowing Zanshin to work with an external system. Plus, major refactoring in the bundles. Read below.
1. Project zanshin-remote was created to implement the API that the
target system will use to communicate with Zanshin.

2. Project zanshin-simulations was created to implement the simulations
that were originally in the it.unitn.disi.zanshin.simulation bundle. The
idea is to have the simulations as external systems too, to serve as an
example to anyone who wants to apply Zanshin to a real system.

3. Project it.unitn.disi.zanshin.controller was created to implement the
communication to external systems on Zanshin's side. It registers an RMI
server that is contacted by the zanshin-remote API.

4. Dependencies among bundles were changed in all projects. Before, the
dependency was to the precise bundle. Now the dependencies are to the
packages, which means that other bundles offering the same packages can
be used instead.

Zanshin's controller already implements part of the system registration
feature, which is the first thing an external system does when talking
to Zanshin. Zanshin receives the system's  requirements as an EMF file
and should generate Java sources based on it, compile them and load the
classes in its own classpath, so that the monitoring component will be
notified when life-cycle methods of these requirements are called. 

Java source generation is done. To compile the classes, the classes from
the base packages (eca, gore, LTL) should be added to the project that
is created to hold the generated classes and then the batch compiler
should be called. These should be done next.