Continued the development of Zanshin Remote API.
Fixed the management of models to use the correct packages instead of
generic EPackage instances. Included the submission of the requirements
model as well as the meta-model and the instantiation of goal models in
memory, registering them within the Repository Service. The
zanshin-simulations project now tests the execution of a life-cycle
method in a requirement instance and Zanshin properly captures that in
the monitoring service.

Next, I should port the rest of the Acad simulations from the
simulations bundle to the zanshin-simulations external project, but
still using a mock monitoring service. If the simulations can be
properly run like this, then the mock monitoring services should be
replaced by proper rules monitoring, which means the Zanshin RMI Server
should also receive AwReqs definitions in OCLtm and compile them to a
rules file...