Continued the development of Zanshin's Remote API.
The controller bundle of Zanshin now provides an implementation for the
target system controller service which works together with the Zanshin
RMI Server to send adaptation instructions via RMI to the external
application. The basis of this has been committed here, but simulation 1
is still not complete: the RMI client at the target system's (i.e., the
simulation's) end should "wake the simulation up" after receiving all
the adaptation instructions. Will do that next, and then implement the
other simulations.

Note that the simulations bundle has been removed from the project. Most
of its classes will come back, though, as part of the
zanshin-simulations project.