Removed aspects and simplified the launch configuration.
Since the ECA EMF classes already had code that retrieved the Target
System Controller Service from the Activator of the core bundle,
replaced the monitoring aspect with a direct call to the Monitoring
Service, obtaining a reference to it from the Activator.

Moreover, added the required bundles for using the OSGi Console as
direct dependencies of the core bundle, in order to simplify the launch
configuration for Zanshin.